+++ README.md
@@ -2,6 +2,16 @@
 
 comparejs implements JavaScript's comparison operators the way you would expect them to be.
 
+## Status
+
+| Category         | Status                                                                                                                                         |
+| ---------------- | ---------------------------------------------------------------------------------------------------------------------------------------------- |
+| Version          | [![npm](https://img.shields.io/npm/v/comparejs)](https://www.npmjs.com/package/comparejs)                                                      |
+| Dependencies     | ![David](https://img.shields.io/david/thenativeweb/comparejs)                                                                                  |
+| Dev dependencies | ![David](https://img.shields.io/david/dev/thenativeweb/comparejs)                                                                              |
+| Build            | [![CircleCI](https://img.shields.io/circleci/build/github/thenativeweb/comparejs)](https://circleci.com/gh/thenativeweb/comparejs/tree/master) |
+| License          | ![GitHub](https://img.shields.io/github/license/thenativeweb/comparejs)                                                                        |
+
 ## Installation
 
 ```shell
@@ -13,19 +23,23 @@ $ npm install comparejs
 To use comparejs you need to reference it from your application by adding the following line:
 
 ```javascript
-const compare = require('comparejs');
+const compare = require('comparejs').default;
+```
+
+```typescript
+import compare from 'comparejs';
 ```
 
 ### Features
 
--   Supports comparison of `array`, `boolean`, `function`, `null`, `number`, `object`, `string` and `undefined`.
--   Supports comparison of native data types and constructor-created data types, such as `number` and `new Number()`.
--   Supports comparison of objects and arrays using deep-equal.
--   Supports comparison of objects and arrays with `<`, `<=`, `>` and `>=` by handling them as subsets.
--   Supports comparison of objects by structure.
--   Supports comparison with `undefined` correctly, as `<=` and `>=` are problematic in plain JavaScript.
--   Supports comparison in a perfectly type-safe way out-of-the-box.
--   Supports comparison by equality and identity, depending on what makes sense.
+- Supports comparison of `array`, `boolean`, `function`, `null`, `number`, `object`, `string` and `undefined`.
+- Supports comparison of native data types and constructor-created data types, such as `number` and `new Number()`.
+- Supports comparison of objects and arrays using deep-equal.
+- Supports comparison of objects and arrays with `<`, `<=`, `>` and `>=` by handling them as subsets.
+- Supports comparison of objects by structure.
+- Supports comparison with `undefined` correctly, as `<=` and `>=` are problematic in plain JavaScript.
+- Supports comparison in a perfectly type-safe way out-of-the-box.
+- Supports comparison by equality and identity, depending on what makes sense.
 
 ### Basic usage
 
@@ -65,14 +79,3 @@ To build this module use [roboter](https://www.npmjs.com/package/roboter).
 ```shell
 $ npx roboter
 ```
-
-## License
-
-The MIT License (MIT)
-Copyright (c) 2012-2019 the native web.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
