+++ lib/unbox.ts
@@ -1,6 +1,5 @@
-'use strict';
-
-const unbox = function (value) {
+/* eslint-disable @typescript-eslint/promise-function-async */
+const unbox = function (value: any): any {
   if (typeof value !== 'object') {
     return value;
   }
@@ -16,5 +15,6 @@ const unbox = function (value) {
 
   return value.valueOf();
 };
+/* eslint-enable @typescript-eslint/promise-function-async */
 
-module.exports = unbox;
+export default unbox;
