+++ test/unit/circularTests.ts
@@ -1,16 +1,12 @@
-'use strict';
+import cmp from '../../lib/compare';
+import 'should';
 
-/* eslint-disable no-unused-vars */
-const should = require('should');
-/* eslint-enable no-unused-vars */
-
-const cmp = require('../../lib/compare');
-
-suite('cmp', () => {
-  suite('eq', () => {
-    test('first with self reference => false', done => {
+suite('cmp', (): void => {
+  suite('eq', (): void => {
+    test('first with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -19,43 +15,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.eq(first, second).should.equal(false);
-      done();
     });
 
-    test('second with self reference => false', done => {
+    test('second with self reference => false', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.eq(first, second).should.equal(false);
-      done();
     });
 
-    test('both with self reference => true', done => {
+    test('both with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.eq(first, second).should.equal(true);
-      done();
     });
   });
 
-  suite('eqs', () => {
-    test('first with self reference => false', done => {
+  suite('eqs', (): void => {
+    test('first with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -64,43 +61,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.eqs(first, second).should.equal(false);
-      done();
     });
 
-    test('second with self reference => false', done => {
+    test('second with self reference => false', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.eqs(first, second).should.equal(false);
-      done();
     });
 
-    test('both with self reference => true', done => {
+    test('both with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.eqs(first, second).should.equal(true);
-      done();
     });
   });
 
-  suite('ne', () => {
-    test('first with self reference => true', done => {
+  suite('ne', (): void => {
+    test('first with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -109,43 +107,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.ne(first, second).should.equal(true);
-      done();
     });
 
-    test('second with self reference => true', done => {
+    test('second with self reference => true', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.ne(first, second).should.equal(true);
-      done();
     });
 
-    test('both with self reference => false', done => {
+    test('both with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.ne(first, second).should.equal(false);
-      done();
     });
   });
 
-  suite('nes', () => {
-    test('first with self reference => true', done => {
+  suite('nes', (): void => {
+    test('first with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -154,43 +153,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.nes(first, second).should.equal(true);
-      done();
     });
 
-    test('second with self reference => true', done => {
+    test('second with self reference => true', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.nes(first, second).should.equal(true);
-      done();
     });
 
-    test('both with self reference => false', done => {
+    test('both with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.nes(first, second).should.equal(false);
-      done();
     });
   });
 
-  suite('gt', () => {
-    test('first with self reference => true', done => {
+  suite('gt', (): void => {
+    test('first with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -199,43 +199,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.gt(first, second).should.equal(true);
-      done();
     });
 
-    test('second with self reference => false', done => {
+    test('second with self reference => false', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.gt(first, second).should.equal(false);
-      done();
     });
 
-    test('both with self reference => false', done => {
+    test('both with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.gt(first, second).should.equal(false);
-      done();
     });
   });
 
-  suite('gts', () => {
-    test('first with self reference => true', done => {
+  suite('gts', (): void => {
+    test('first with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -244,43 +245,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.gts(first, second).should.equal(true);
-      done();
     });
 
-    test('second with self reference => false', done => {
+    test('second with self reference => false', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.gts(first, second).should.equal(false);
-      done();
     });
 
-    test('both with self reference => false', done => {
+    test('both with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.gts(first, second).should.equal(false);
-      done();
     });
   });
 
-  suite('gte', () => {
-    test('first with self reference => true', done => {
+  suite('gte', (): void => {
+    test('first with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -289,43 +291,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.gte(first, second).should.equal(true);
-      done();
     });
 
-    test('second with self reference => false', done => {
+    test('second with self reference => false', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.gte(first, second).should.equal(false);
-      done();
     });
 
-    test('both with self reference => true', done => {
+    test('both with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.gte(first, second).should.equal(true);
-      done();
     });
   });
 
-  suite('gtes', () => {
-    test('first with self reference => true', done => {
+  suite('gtes', (): void => {
+    test('first with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -334,43 +337,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.gtes(first, second).should.equal(true);
-      done();
     });
 
-    test('second with self reference => false', done => {
+    test('second with self reference => false', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.gtes(first, second).should.equal(false);
-      done();
     });
 
-    test('both with self reference => true', done => {
+    test('both with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.gtes(first, second).should.equal(true);
-      done();
     });
   });
 
-  suite('lt', () => {
-    test('first with self reference => false', done => {
+  suite('lt', (): void => {
+    test('first with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -379,43 +383,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.lt(first, second).should.equal(false);
-      done();
     });
 
-    test('second with self reference => true', done => {
+    test('second with self reference => true', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.lt(first, second).should.equal(true);
-      done();
     });
 
-    test('both with self reference => false', done => {
+    test('both with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.lt(first, second).should.equal(false);
-      done();
     });
   });
 
-  suite('lts', () => {
-    test('first with self reference => false', done => {
+  suite('lts', (): void => {
+    test('first with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -424,43 +429,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.lts(first, second).should.equal(false);
-      done();
     });
 
-    test('second with self reference => true', done => {
+    test('second with self reference => true', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.lts(first, second).should.equal(true);
-      done();
     });
 
-    test('both with self reference => false', done => {
+    test('both with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.lts(first, second).should.equal(false);
-      done();
     });
   });
 
-  suite('lte', () => {
-    test('first with self reference => false', done => {
+  suite('lte', (): void => {
+    test('first with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -469,43 +475,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.lte(first, second).should.equal(false);
-      done();
     });
 
-    test('second with self reference => true', done => {
+    test('second with self reference => true', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.lte(first, second).should.equal(true);
-      done();
     });
 
-    test('both with self reference => true', done => {
+    test('both with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.lte(first, second).should.equal(true);
-      done();
     });
   });
 
-  suite('ltes', () => {
-    test('first with self reference => false', done => {
+  suite('ltes', (): void => {
+    test('first with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -514,43 +521,44 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.ltes(first, second).should.equal(false);
-      done();
     });
 
-    test('second with self reference => true', done => {
+    test('second with self reference => true', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.ltes(first, second).should.equal(true);
-      done();
     });
 
-    test('both with self reference => true', done => {
+    test('both with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.ltes(first, second).should.equal(true);
-      done();
     });
   });
 
-  suite('id', () => {
-    test('first with self reference => false', done => {
+  suite('id', (): void => {
+    test('first with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
@@ -559,46 +567,46 @@ suite('cmp', () => {
 
       first.mySelf = first;
       cmp.id(first, second).should.equal(false);
-      done();
     });
 
-    test('second with self reference => false', done => {
+    test('second with self reference => false', async (): Promise<void> => {
       const first = {
         name: 'foo'
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       second.mySelf = second;
       cmp.id(first, second).should.equal(false);
-      done();
     });
 
-    test('both with self reference => false', done => {
+    test('both with self reference => false', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       const second = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       second.mySelf = second;
       cmp.id(first, second).should.equal(false);
-      done();
     });
 
-    test('identical with self reference => true', done => {
+    test('identical with self reference => true', async (): Promise<void> => {
       const first = {
-        name: 'foo'
+        name: 'foo',
+        mySelf: undefined as any
       };
 
       first.mySelf = first;
       cmp.id(first, first).should.equal(true);
-      done();
     });
   });
 });
